[OI] fix variable rendering influenced by "upper" filter
According to
https://manpages.debian.org/testing/auditd/auditd.conf.5.en.html, `All
option names and values are case insensitive.`.

However, CIS-CAT scan results fail when values are uppercase. This
is the sole motivation for this change. Clearly, CIS config assessment
tool needs to be fixed. But I suspect it'll be far quicker and easier to
make option values lowercase since that's valid too.

Please consider this change (which doesn't add any "real" value). ;-)